feat: extracted actions to separate files, update references for old message model
updates to model

fix conflicts

Updated model
