+++ Susi/Controllers/LoginController/LoginViewController.swift
@@ -11,6 +11,7 @@ import Material
 import Toast_Swift
 import SwiftValidators
 import DLRadioButton
+import RealmSwift
 
 class LoginViewController: UIViewController {
 
@@ -105,6 +106,8 @@ class LoginViewController: UIViewController {
         setupView()
         checkSession()
         addTapGesture()
+
+        print(Realm.Configuration.defaultConfiguration.fileURL!)
     }
 
 }
