+++ Susi/Helpers/ControllerConstants.swift
@@ -84,6 +84,7 @@ class ControllerConstants {
         static let speechOutputAlwaysOn = "speechOutputAlwaysOn"
         static let wallpaper = "wallpaper"
         static let ipAddress = "ipAddress"
+        static let firstLaunch = "firstLaunch"
     }
 
 }
