+++ Susi/Model/AnchorAction.swift
@@ -1,0 +1,25 @@
+//
+//  AnchorAction.swift
+//  Susi
+//
+//  Created by Chashmeet Singh on 2017-06-10.
+//  Copyright © 2017 FOSSAsia. All rights reserved.
+//
+
+import Foundation
+import RealmSwift
+
+class AnchorAction: Message {
+    dynamic var link: String = ""
+    dynamic var text: String = ""
+
+    convenience init(data: [String : AnyObject]) {
+        self.init()
+
+        if let link = data[Client.ChatKeys.Link] as? String,
+            let text = data[Client.ChatKeys.Text] as? String {
+            self.link = link
+            self.text = text
+        }
+    }
+}
