+++ Susi/Model/Message.swift
@@ -2,108 +2,96 @@
 //  Message.swift
 //  Susi
 //
-//  Created by Chashmeet Singh on 2017-03-13.
+//  Created by Chashmeet Singh on 2017-06-10.
 //  Copyright © 2017 FOSSAsia. All rights reserved.
 //
 
-import UIKit
+import Foundation
+import RealmSwift
 
-struct Message {
+enum ActionType: String {
+    case answer
+    case websearch
+    case piechart
+    case rss
+    case table
+    case map
+    case anchor
+}
 
-    var body: String? = ""
-    var created_at = Date()
-    var isBot = false
-    var responseType: ResponseTypes = .answer
+class Message: Object {
+    dynamic var queryDate = NSDate()
+    dynamic var answerDate = NSDate()
+    dynamic var query: String = ""
+    dynamic var isSent = false
+    dynamic var isMarked = false
+    dynamic var fromUser = true
+    let messages = List<Message>()
 
-    var mapData: MapData?
-
-    struct MapData {
-        var longitude: Double
-        var latitude: Double
-        var zoom: Int
+    convenience init(messageBody: String) {
+        self.init()
+        self.query = messageBody
     }
 
-    var websearchData: WebsearchResult?
-    var query: String?
+    convenience init(dictionary: [String : AnyObject]) {
+        self.init()
 
-    enum ResponseTypes: String {
-        case answer = "answer"
-        case map = "map"
-        case websearch = "websearch"
-        case image = "image"
-    }
+        self.fromUser = false
 
-    init(_ body: String) {
-        self.body = body
-    }
-
-    init(dictionary: [String : AnyObject], isBot: Bool) {
-
-        if let created_at = dictionary[Client.ChatKeys.AnswerDate] as? String {
-            let dateFormatter = DateFormatter()
-            dateFormatter.locale = NSLocale(localeIdentifier: "en_US_POSIX") as Locale!
-            dateFormatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ss.SSS'Z"
-
-            self.created_at = dateFormatter.date(from: created_at)!
+        if let query = dictionary[Client.ChatKeys.Query] as? String {
+            self.query = query
+        }
+        
+        let dateFormatter = DateFormatter()
+        dateFormatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ss.SSSZ"
+        
+        if let queryDate = dictionary[Client.ChatKeys.QueryDate] as? String,
+            let answerDate = dictionary[Client.ChatKeys.AnswerDate] as? String {
+            self.queryDate = dateFormatter.date(from: queryDate)! as NSDate
+            self.answerDate = dateFormatter.date(from: answerDate)! as NSDate
         }
 
-        if let body = dictionary[Client.ChatKeys.Answers] as? [[String : AnyObject]] {
-            if body.count > 0 {
-                if let actions = body[0][Client.ChatKeys.Actions] as? NSArray {
-                    if let response = actions[0] as? [String : String] {
-                        self.body = response[Client.ChatKeys.Expression]!
-                    }
-
-                    if let responses = actions as? [[String : String]] {
-
-                        for response in responses {
-                            let responseType = response[Client.ChatKeys.ResponseType]
-
-                            if responseType == ResponseTypes.answer.rawValue {
-
-                                if let expression = response[Client.ChatKeys.Expression] {
-                                    if expression.isURL() && expression.isImage() {
-                                        self.responseType = ResponseTypes.image
-                                    }
-                                    self.body = expression
-                                }
-
-                            } else if responseType == ResponseTypes.map.rawValue {
-
-                                self.responseType = ResponseTypes.map
-
-                                if let latitude = response[Client.ChatKeys.Latitude],
-                                    let longitude = response[Client.ChatKeys.Longitude],
-                                    let zoom = response[Client.ChatKeys.Zoom] {
-
-                                    self.mapData = MapData(longitude: Double(longitude)!, latitude: Double(latitude)!, zoom: Int(zoom)!)
-                                }
-
-                                self.body = self.body?.components(separatedBy: " ").dropLast().joined(separator: " ")
-
-                            } else if responseType == ResponseTypes.websearch.rawValue {
-
-                                self.responseType = ResponseTypes.websearch
-                                self.query = response[Client.ChatKeys.Query]
-
-                            } else {
-                                debugPrint("error")
-                            }
+        if let answers = dictionary[Client.ChatKeys.Answers] as? [[String : AnyObject]] {
+            if let actions = answers[0][Client.ChatKeys.Actions] as? [[String : AnyObject]] {
+                for action in actions {
+                    if let type = action[Client.ChatKeys.ResponseType] as? String,
+                        let data = answers[0][Client.ChatKeys.Data] as? [[String : AnyObject]] {
+                        if type == ActionType.answer.rawValue {
+                            let answer = AnswerAction(data: action)
+                            answer.queryDate = self.queryDate
+                            answer.answerDate = self.answerDate
+                            answer.query = self.query
+                            self.messages.append(answer)
+                        } else if type == ActionType.rss.rawValue {
+                            let rss = RSSAction(data: data, actionObject: action)
+                            rss.queryDate = self.queryDate
+                            rss.answerDate = self.answerDate
+                            rss.query = self.query
+                            self.messages.append(rss)
+                        } else if type == ActionType.map.rawValue {
+                            let map = MapAction(data: action)
+                            map.queryDate = self.queryDate
+                            map.answerDate = self.answerDate
+                            map.query = self.query
+                            self.messages.append(map)
+                        } else if type == ActionType.anchor.rawValue {
+                            let anchor = AnchorAction(data: action)
+                            anchor.queryDate = self.queryDate
+                            anchor.answerDate = self.answerDate
+                            anchor.query = self.query
+                            self.messages.append(anchor)
+                        } else if type == ActionType.table.rawValue {
+                            let table = TableAction(data: data, actionObject: action)
+                            table.queryDate = self.queryDate
+                            table.answerDate = self.answerDate
+                            table.query = self.query
+                            self.messages.append(table)
                         }
                     }
                 }
-            } else {
-                debugPrint(body)
             }
         }
 
-        self.isBot = isBot
-    }
-
-    static func getMessageFromResponse(_ result: [String : AnyObject], isBot: Bool) -> Message {
-        let message = Message(dictionary: result, isBot: isBot)
-
-        return message
     }
 
 }
